updated superimpose.py to accept index files
