+++ superimpose.py
@@ -10,48 +10,61 @@ try:
 except:
     from bio.SVDSuperimposer import SVDSuperimposer
 from sys import exit, stderr
-from UTILS.readers import LorenzoReader2
+from UTILS.readers import ErikReader
 import numpy as np
 import argparse
 from compute_mean import normalize
 
-fetch_np = lambda conf: np.array([
-    n.cm_pos for n in conf._nucleotides 
-])
-
 parser = argparse.ArgumentParser(description="superimposes one or more structures sharing a topology to a reference structure")
-parser.add_argument('topology', type=str, nargs=1, help="The topology file shared by all the configurations to superimpose")
 parser.add_argument('reference', type=str, nargs=1, help="The reference configuration to superimpose to")
 parser.add_argument('victims', type=str, nargs='+', help="The configuraitons to superimpose on the reference")
+parser.add_argument('-i', metavar='index_file', dest='index_file', nargs=1, help='Align to only a subset of particles from a space-separated list in the provided file')
 args = parser.parse_args()
 
+
+#run system checks
+from config import check_dependencies
+check_dependencies(["python", "numpy", "Bio"])
+
 #Get the reference files
-top_file = args.topology[0]
 ref_dat = args.reference[0]
+
+#-i will make it only run on a subset of nucleotides.
+#The index file is a space-separated list of particle IDs
+if args.index_file:
+    index_file = args.index_file[0]
+    with open(index_file, 'r') as f:
+        indexes = f.readline().split()
+        try:
+            indexes = [int(i) for i in indexes]
+        except:
+            print("ERROR: The index file must be a space-seperated list of particles.  These can be generated using oxView by clicking the \"Download Selected Base List\" button")
+else: 
+    with ErikReader(ref_dat) as r:
+        indexes = list(range(len(r.read().positions)))
 
 #Create list of configurations to superimpose
 to_sup = []
-r = LorenzoReader2(ref_dat, top_file)
-ref = r._get_system()
+r = ErikReader(ref_dat)
+ref = r.read()
 ref.inbox()
-ref_conf = fetch_np(ref)
+ref_conf = ref.positions[indexes]
 for i in args.victims:
-    r = LorenzoReader2(i, top_file)
-    sys = r._get_system()
+    r = ErikReader(i)
+    sys = r.read()
+    sys.inbox()
     to_sup.append(sys)
 
 sup = SVDSuperimposer()
 
 #Run the biopython superimposer on each configuration and rewrite its configuration file
 for i, sys in enumerate(to_sup):
-    cur_conf = fetch_np(sys)
-    sup.set(ref_conf, cur_conf)
+    indexed_cur_conf = sys.positions[indexes]
+    sup.set(ref_conf, indexed_cur_conf)
     sup.run()
     rot, tran = sup.get_rotran()
-    cur_conf = np.einsum('ij, ki -> kj', rot, cur_conf) + tran
-    for j,n in enumerate(sys._nucleotides):
-        n.cm_pos = cur_conf[j]
-        n._a1 = normalize(np.dot(n._a1, rot))
-        n._a3 = normalize(np.dot(n._a3, rot))
-    sys.print_lorenzo_output("aligned{}.dat".format(i), "/dev/null")
+    sys.positions = np.einsum('ij, ki -> kj', rot, sys.positions) + tran
+    sys.a1s = np.einsum('ij, ki -> kj', rot, sys.a1s)
+    sys.a3s = np.einsum('ij, ki -> kj', rot, sys.a3s)
+    sys.write_new("aligned{}.dat".format(i))
     print("INFO: Wrote file aligned{}.dat".format(i), file=stderr)
